% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{Database}
\alias{Database}
\title{Database}
\description{
Database

Database

Intitialize
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{msl}}{table: master security list}

\item{\code{geo}}{table: country number and geography}

\item{\code{api_keys}}{optional list containing api_keys}

\item{\code{bucket}}{`S3FileSystem` object from `arrow`}

\item{\code{macro}}{list with tables for R3000 and ACWI model from Piper Sandler}

\item{\code{ret}}{list with daily and monthly return tables}

\item{\code{api_file}}{optional .RData file to load list of api_keys}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{api_file}}{optional .RData file to load list of api_keys}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Database-new}{\code{Database$new()}}
\item \href{#method-Database-write_msl}{\code{Database$write_msl()}}
\item \href{#method-Database-read_msl}{\code{Database$read_msl()}}
\item \href{#method-Database-read_msl_xl}{\code{Database$read_msl_xl()}}
\item \href{#method-Database-write_geo}{\code{Database$write_geo()}}
\item \href{#method-Database-read_geo}{\code{Database$read_geo()}}
\item \href{#method-Database-read_geo_xl}{\code{Database$read_geo_xl()}}
\item \href{#method-Database-read_macro}{\code{Database$read_macro()}}
\item \href{#method-Database-check_bucket}{\code{Database$check_bucket()}}
\item \href{#method-Database-check_msl}{\code{Database$check_msl()}}
\item \href{#method-Database-check_geo}{\code{Database$check_geo()}}
\item \href{#method-Database-check_api_keys}{\code{Database$check_api_keys()}}
\item \href{#method-Database-update_holding}{\code{Database$update_holding()}}
\item \href{#method-Database-update_all_ctf_holdings}{\code{Database$update_all_ctf_holdings()}}
\item \href{#method-Database-update_ctf_holdings}{\code{Database$update_ctf_holdings()}}
\item \href{#method-Database-update_sec}{\code{Database$update_sec()}}
\item \href{#method-Database-update_bd}{\code{Database$update_bd()}}
\item \href{#method-Database-update_all_port}{\code{Database$update_all_port()}}
\item \href{#method-Database-get_curr_holdings}{\code{Database$get_curr_holdings()}}
\item \href{#method-Database-read_all_ret}{\code{Database$read_all_ret()}}
\item \href{#method-Database-filter_fs_ids}{\code{Database$filter_fs_ids()}}
\item \href{#method-Database-update_fs_fina_most_recent}{\code{Database$update_fs_fina_most_recent()}}
\item \href{#method-Database-update_fs_fina_quarterly}{\code{Database$update_fs_fina_quarterly()}}
\item \href{#method-Database-update_fs_mf_ret_daily}{\code{Database$update_fs_mf_ret_daily()}}
\item \href{#method-Database-update_fs_ret_daily}{\code{Database$update_fs_ret_daily()}}
\item \href{#method-Database-update_ctf_daily}{\code{Database$update_ctf_daily()}}
\item \href{#method-Database-update_pdf_funds}{\code{Database$update_pdf_funds()}}
\item \href{#method-Database-load_ret}{\code{Database$load_ret()}}
\item \href{#method-Database-update_macro}{\code{Database$update_macro()}}
\item \href{#method-Database-missing_from_msl}{\code{Database$missing_from_msl()}}
\item \href{#method-Database-check_ctf_holdings}{\code{Database$check_ctf_holdings()}}
\item \href{#method-Database-clone}{\code{Database$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-new"></a>}}
\if{latex}{\out{\hypertarget{method-Database-new}{}}}
\subsection{Method \code{new()}}{
Create a Database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$new(
  msl = NULL,
  geo = NULL,
  bucket = NULL,
  api_keys = NULL,
  api_file = "N:/Investment Team/DATABASES/MDB/Keys/api_keys.RData"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msl}}{table: master security list}

\item{\code{geo}}{table: country number and geography}

\item{\code{bucket}}{`S3FileSystem` object from `arrow`}

\item{\code{api_keys}}{list of api_keys}

\item{\code{api_file}}{`.RData` file to load api_key list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
All fields are optional, except the api_keys must be either entered
as a list `api_keys` or a file path to load the api_keys `api_file`.
The tables are `data.frames` or `tibbles` stored as arrow or parquet
files in S3, accessed through the `bucket` or `S3FileSystem`.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-write_msl"></a>}}
\if{latex}{\out{\hypertarget{method-Database-write_msl}{}}}
\subsection{Method \code{write_msl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$write_msl(msl = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-read_msl"></a>}}
\if{latex}{\out{\hypertarget{method-Database-read_msl}{}}}
\subsection{Method \code{read_msl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$read_msl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-read_msl_xl"></a>}}
\if{latex}{\out{\hypertarget{method-Database-read_msl_xl}{}}}
\subsection{Method \code{read_msl_xl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$read_msl_xl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-write_geo"></a>}}
\if{latex}{\out{\hypertarget{method-Database-write_geo}{}}}
\subsection{Method \code{write_geo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$write_geo(geo = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-read_geo"></a>}}
\if{latex}{\out{\hypertarget{method-Database-read_geo}{}}}
\subsection{Method \code{read_geo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$read_geo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-read_geo_xl"></a>}}
\if{latex}{\out{\hypertarget{method-Database-read_geo_xl}{}}}
\subsection{Method \code{read_geo_xl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$read_geo_xl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-read_macro"></a>}}
\if{latex}{\out{\hypertarget{method-Database-read_macro}{}}}
\subsection{Method \code{read_macro()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$read_macro()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-check_bucket"></a>}}
\if{latex}{\out{\hypertarget{method-Database-check_bucket}{}}}
\subsection{Method \code{check_bucket()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$check_bucket()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-check_msl"></a>}}
\if{latex}{\out{\hypertarget{method-Database-check_msl}{}}}
\subsection{Method \code{check_msl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$check_msl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-check_geo"></a>}}
\if{latex}{\out{\hypertarget{method-Database-check_geo}{}}}
\subsection{Method \code{check_geo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$check_geo()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-check_api_keys"></a>}}
\if{latex}{\out{\hypertarget{method-Database-check_api_keys}{}}}
\subsection{Method \code{check_api_keys()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$check_api_keys()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_holding"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_holding}{}}}
\subsection{Method \code{update_holding()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_holding(
  dtc_name,
  as_of = NULL,
  user_email = NULL,
  xl_df = NULL,
  return_df = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_all_ctf_holdings"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_all_ctf_holdings}{}}}
\subsection{Method \code{update_all_ctf_holdings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_all_ctf_holdings(update_bd = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_ctf_holdings"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_ctf_holdings}{}}}
\subsection{Method \code{update_ctf_holdings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_ctf_holdings(dtc_name, add_to_existing = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_sec"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_sec}{}}}
\subsection{Method \code{update_sec()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_sec(obs, user_email)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_bd"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_bd}{}}}
\subsection{Method \code{update_bd()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_bd(obs, api_keys, as_of)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_all_port"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_all_port}{}}}
\subsection{Method \code{update_all_port()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_all_port(user_email = NULL, as_of = NULL, update_csv = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-get_curr_holdings"></a>}}
\if{latex}{\out{\hypertarget{method-Database-get_curr_holdings}{}}}
\subsection{Method \code{get_curr_holdings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$get_curr_holdings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-read_all_ret"></a>}}
\if{latex}{\out{\hypertarget{method-Database-read_all_ret}{}}}
\subsection{Method \code{read_all_ret()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$read_all_ret()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-filter_fs_ids"></a>}}
\if{latex}{\out{\hypertarget{method-Database-filter_fs_ids}{}}}
\subsection{Method \code{filter_fs_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$filter_fs_ids(max_iter_by = 50)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_fs_fina_most_recent"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_fs_fina_most_recent}{}}}
\subsection{Method \code{update_fs_fina_most_recent()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_fs_fina_most_recent()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_fs_fina_quarterly"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_fs_fina_quarterly}{}}}
\subsection{Method \code{update_fs_fina_quarterly()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_fs_fina_quarterly(
  yrs_back = 1,
  dtype = c("PE", "PB", "PFCF", "DY", "ROE", "MCAP")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_fs_mf_ret_daily"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_fs_mf_ret_daily}{}}}
\subsection{Method \code{update_fs_mf_ret_daily()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_fs_mf_ret_daily(days_back = 0)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_fs_ret_daily"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_fs_ret_daily}{}}}
\subsection{Method \code{update_fs_ret_daily()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_fs_ret_daily(ids = NULL, date_start = NULL, date_end = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_ctf_daily"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_ctf_daily}{}}}
\subsection{Method \code{update_ctf_daily()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_ctf_daily()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_pdf_funds"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_pdf_funds}{}}}
\subsection{Method \code{update_pdf_funds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_pdf_funds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-load_ret"></a>}}
\if{latex}{\out{\hypertarget{method-Database-load_ret}{}}}
\subsection{Method \code{load_ret()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$load_ret()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-update_macro"></a>}}
\if{latex}{\out{\hypertarget{method-Database-update_macro}{}}}
\subsection{Method \code{update_macro()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$update_macro(
  fpath = NULL,
  acwi = "D_MACRO_SELECT_GLOBAL",
  r3 = "D_MACRO_SELECT_R",
  fct_name = c("A", "B", "C", "D", "E")
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-missing_from_msl"></a>}}
\if{latex}{\out{\hypertarget{method-Database-missing_from_msl}{}}}
\subsection{Method \code{missing_from_msl()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$missing_from_msl()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-check_ctf_holdings"></a>}}
\if{latex}{\out{\hypertarget{method-Database-check_ctf_holdings}{}}}
\subsection{Method \code{check_ctf_holdings()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$check_ctf_holdings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Database-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Database-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Database$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
