% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_from_source.R
\name{download_fs_gp}
\alias{download_fs_gp}
\title{Download Factset Global Prices}
\usage{
download_fs_gp(api_keys, ids, date_start, date_end, freq = "D")
}
\arguments{
\item{api_keys}{list with factset api keys}

\item{ids}{requested security identifiers, can by tickers, SEDOL, ISIN, 
CUSIP, or FactSet Permanent Id}

\item{date_start}{string in YYYY-MM-DD format}

\item{date_end}{same}

\item{freq}{string, "D" = daily}
}
\value{
json object of total returns
}
\description{
Download Factset Global Prices
}
