% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_from_source.R
\name{download_fs_large_names}
\alias{download_fs_large_names}
\title{Download company name of from large number of ids}
\usage{
download_fs_large_names(api_keys, ids)
}
\arguments{
\item{api_keys}{list with api_keys, see Database}

\item{ids}{requestIds}
}
\description{
Download company name of from large number of ids
}
