% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_from_source.R
\name{download_tiingo_csv}
\alias{download_tiingo_csv}
\title{Download price time-series from Tiingo}
\usage{
download_tiingo_csv(ticker, t_api, start_date = "1970-01-01", end_date = NULL)
}
\arguments{
\item{ticker}{character of stock ticker to download, see
`download_tiingo_tickers` for downloading multiple tickers}

\item{t_api}{API key}

\item{start_date}{inception for time-series}

\item{end_date}{last day of time-series, if left `NULL` will default to today}
}
\value{
`data.frame` of historical prices (adjusted for splits and dividends)
  with dates in first column
}
\description{
Download price time-series from Tiingo
}
