% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_from_source.R
\name{download_tiingo_tickers}
\alias{download_tiingo_tickers}
\title{Download multiple tickers from Tiingo}
\usage{
download_tiingo_tickers(ticker_vec, t_api, start_date, end_date = NULL)
}
\arguments{
\item{ticker_vec}{character vector of tickers to download}

\item{t_api}{API key}

\item{start_date}{first date in time-series to download}

\item{end_date}{last date in time-series to download, if left `NULL` will
default to today}
}
\value{
xts of price time-series adjusted for dividends and splits
}
\description{
Download multiple tickers from Tiingo
}
\examples{
\dontrun {
  download_tiingo_tickers(c('AAPL', 'GOOG', 'TSLA'), api_key, '2020-01-01')
}
}
