% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_from_source.R
\name{flatten_fs_gp}
\alias{flatten_fs_gp}
\title{Convert json output of factset global prices into data.frame}
\usage{
flatten_fs_gp(json)
}
\arguments{
\item{json}{download from `download_fs_gp` function}
}
\description{
Convert json output of factset global prices into data.frame
}
\note{
see download_fs_gp
}
