% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_from_source.R
\name{unlist_fs_ts}
\alias{unlist_fs_ts}
\title{Transform FactSet time-series JSON into data.frame}
\usage{
unlist_fs_ts(json)
}
\arguments{
\item{json}{json list output from `download_fs`}
}
\value{
data.frame
}
\description{
Transform FactSet time-series JSON into data.frame
}
